backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
numpy==1.26.4
scipy==1.14.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
